inscripciones-frontend-master/source/src/app/components/baseInputs/select/select.component.ts
import { Component, Input, OnInit } from '@angular/core';
import { FormControl } from '@angular/forms';

@Component({
  selector: 'app-select-input',
  templateUrl: './select.component.html',
  styleUrls: ['./select.component.css'],
})
export class SelectComponent implements OnInit {
  @Input() id: string;
  @Input() control: any = new FormControl();
  @Input() label: string;
  @Input() placeholder: string;
  @Input() disabled: boolean = false;
  @Input() listaOptions: Array<any> = [];
  @Input() name: string;

  constructor() {}

  ngOnInit(): void {}
}
